split StateMachine
