+++ fe/fe-core/src/main/java/com/starrocks/transaction/TransactionLogApplier.java
@@ -1,0 +1,12 @@
+// This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Limited.
+
+package com.starrocks.transaction;
+
+public interface TransactionLogApplier {
+
+    // This method is called by all the FE nodes after sent/received the edit log to change TransactionState to COMMITTED.
+    void applyCommitLog(TransactionState txnState, TableCommitInfo commitInfo);
+
+    // This method is called by all the FE nodes after sent/received the edit log to change the TransactionState to VISIBLE.
+    void applyVisibleLog(TransactionState txnState, TableCommitInfo commitInfo);
+}
