+++ fe/fe-core/src/main/java/com/starrocks/transaction/TransactionLogApplierFactory.java
@@ -1,0 +1,19 @@
+// This file is licensed under the Elastic License 2.0. Copyright 2021-present, StarRocks Limited.
+
+package com.starrocks.transaction;
+
+import com.starrocks.catalog.OlapTable;
+import com.starrocks.catalog.Table;
+import com.starrocks.catalog.lake.LakeTable;
+
+public class TransactionLogApplierFactory {
+    public TransactionLogApplier create(Table table) {
+        if (table.isLakeTable()) {
+            return new LakeTableTxnLogApplier((LakeTable) table);
+        }
+        if (table.isOlapTable()) {
+            return new OlapTableTxnLogApplier((OlapTable) table);
+        }
+        return null;
+    }
+}
