+++ fe/fe-core/src/main/java/com/starrocks/transaction/TransactionStateListener.java
@@ -4,7 +4,7 @@ package com.starrocks.transaction;
 
 import java.util.List;
 
-public interface StateMachine {
+public interface TransactionStateListener {
     // This method is called by the FE master before changing the in-memory TransactionState to COMMITTED.
     void preCommit(TransactionState txnState, List<TabletCommitInfo> finishedTablets) throws TransactionException;
 
@@ -14,10 +14,4 @@ public interface StateMachine {
 
     // This method is called by the FE master after changed the TransactionState to COMMITTED and persisted the edit log.
     void postWriteCommitLog(TransactionState txnState);
-
-    // This method is called by all the FE nodes after sent/received the edit log to change TransactionState to COMMITTED.
-    void applyCommitLog(TransactionState txnState, TableCommitInfo commitInfo);
-
-    // This method is called by all the FE nodes after sent/received the edit log to change the TransactionState to VISIBLE.
-    void applyVisibleLog(TransactionState txnState, TableCommitInfo commitInfo);
 }
